requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
sentence-transformers
faiss-cpu
scikit-fuzzy